moved components from template folder to components folder and updated styling for right container
