+++ src/templates/Dashboard/index.js
@@ -1,7 +1,6 @@
-import { Col } from '@everlywell/leaves';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
-import { LeftNav } from '../LeftNav';
+import { LeftNav } from '../../components/LeftNav';
 import { Home } from '../../pages/Home';
 import { Account } from '../../pages/Account';
 import { HealthProfile } from '../../pages/HealthProfile';
@@ -18,17 +17,15 @@ export const Dashboard = () => {
         <S.LeftColumn>
           <LeftNav />
         </S.LeftColumn>
-        <S.RightColumn style={{'flexBasis': '80%'}}>
-          <S.RightContainer>
-            <Routes>
-              <Route path="/" element={<Home />} />
-              <Route path="/settings" element={<Account />} />
-              <Route path="/health-profile" element={<HealthProfile />} />
-              <Route path="/memberships" element={<Memberships />} />
-              <Route path="/orders" element={<Orders />} />
-              <Route path="/results" element={<Results />} />
-            </Routes>
-          </S.RightContainer>
+        <S.RightColumn>
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/settings" element={<Account />} />
+            <Route path="/health-profile" element={<HealthProfile />} />
+            <Route path="/memberships" element={<Memberships />} />
+            <Route path="/orders" element={<Orders />} />
+            <Route path="/results" element={<Results />} />
+          </Routes>
         </S.RightColumn>
       </S.DashboardContainer>
     </BrowserRouter>
